+++ contracts/LendingPool.sol
@@ -179,8 +179,8 @@ contract LendingPool is Ownable, ILendingPool, IAlphaReceiver {
     bool useAsCollateral;
     // borrow shares of the user of this pool. If user didn't borrow this pool then shere will be 0
     uint256 borrowShares;
-    // latest borrow reward of the suser of this pool. Using to calculate current borrow reward.
-    uint256 latestBorrowReward;
+    // latest alpha multiplier (borrow reward multiplier) of the user of this pool. Using to calculate current borrow reward.
+    uint256 latestAlphaMultiplier;
   }
 
   /**
@@ -204,7 +204,7 @@ contract LendingPool is Ownable, ILendingPool, IAlphaReceiver {
     // total alpha token reward on this pool
     uint256 totalAlphaTokenReward;
     // alpha reward multiplier of each borrow share
-    uint256 alphaRewardMultiplier;
+    uint256 alphaMultiplier;
   }
 
   /**
@@ -779,7 +779,7 @@ contract LendingPool is Ownable, ILendingPool, IAlphaReceiver {
     emit Borrow(address(_token), msg.sender, borrowShare, _amount);
 
     // 6. Update latest borrow reward
-    setLatestReward(_token, msg.sender);
+    setLatestAlphaMultiplier(_token, msg.sender);
   }
 
   /**
@@ -853,7 +853,7 @@ contract LendingPool is Ownable, ILendingPool, IAlphaReceiver {
     emit Repay(address(_token), msg.sender, paybackShares, paybackAmount);
 
     // 5. Update latest borrow reward
-    setLatestReward(_token, msg.sender);
+    setLatestAlphaMultiplier(_token, msg.sender);
   }
 
   /**
@@ -1022,7 +1022,7 @@ contract LendingPool is Ownable, ILendingPool, IAlphaReceiver {
     );
 
     // 12. update latest borrow reward
-    setLatestReward(_token, _user);
+    setLatestAlphaMultiplier(_token, _user);
   }
 
   /**
@@ -1135,16 +1135,25 @@ contract LendingPool is Ownable, ILendingPool, IAlphaReceiver {
     }
   }
 
-  function claimAlphaBorrow() external {
-    uint256 totalReward = 0;
+  /**
+   * @dev claim Alpha token rewards from all ERC20 token pools and create receipt for caller
+   */
+  function claimAlpha() external updateAlphaReward {
+    uint256 totalBorrowReward = 0;
     for (uint256 i = 0; i < tokenList.length; i++) {
-      uint256 reward = calculateAlphaReward(tokenList[i], msg.sender);
-      totalReward = totalReward.add(reward);
       Pool storage pool = pools[address(tokenList[i])];
+
+      // claim Alpha rewards as a lender
+      pool.alToken.claimCurrentAlphaRewardByOwner(msg.sender);
+
+      // claim Alpha rewards a borrower
+      uint256 reward = calculateAlphaReward(tokenList[i], msg.sender);
+      totalBorrowReward = totalBorrowReward.add(reward);
       pool.totalAlphaTokenReward = pool.totalAlphaTokenReward.sub(reward);
-      setLatestReward(tokenList[i], msg.sender);
+      setLatestAlphaMultiplier(tokenList[i], msg.sender);
     }
-    sendAlphaReward(msg.sender, totalReward);
+    // send borrow rewards to vesting contract
+    sendAlphaReward(msg.sender, totalBorrowReward);
   }
 
   /**
@@ -1157,7 +1166,7 @@ contract LendingPool is Ownable, ILendingPool, IAlphaReceiver {
     if (_pool.totalBorrowShares == 0) {
       return;
     }
-    _pool.alphaRewardMultiplier = _pool.alphaRewardMultiplier.add(
+    _pool.alphaMultiplier = _pool.alphaMultiplier.add(
       _amount.mul(1e12).div(_pool.totalBorrowShares)
     );
   }
@@ -1203,12 +1212,22 @@ contract LendingPool is Ownable, ILendingPool, IAlphaReceiver {
   function calculateAlphaReward(ERC20 _token, address _account) public view returns (uint256) {
     Pool storage pool = pools[address(_token)];
     UserPoolData storage userData = userPoolData[_account][address(_token)];
-    uint256 pending = userData.borrowShares.mul(pool.alphaRewardMultiplier).div(1e12).sub(
-      userData.latestBorrowReward
-    );
+    //               reward start block                                        now
+    // Global                |----------------|----------------|----------------|
+    // User's latest reward  |----------------|----------------|
+    // User's Alpha rewards                                    |----------------|
+    // reward = [(Global Alpha multiplier - user's lastest Alpha multiplier) * user's Alpha token] / 1e12
+    uint256 pending = pool
+      .alphaMultiplier
+      .sub(userData.latestAlphaMultiplier)
+      .mul(userData.borrowShares)
+      .div(1e12);
     return pending < pool.totalAlphaTokenReward ? pending : pool.totalAlphaTokenReward;
   }
 
+   /**
+   * @dev claim Alpha tokens rewards
+   */
   function claimCurrentAlphaReward(ERC20 _token, address _account) internal {
     // No op if alpha distributor didn't be set in lending pool.
     if (address(distributor) == address(0)) {
@@ -1218,9 +1237,12 @@ contract LendingPool is Ownable, ILendingPool, IAlphaReceiver {
     uint256 reward = calculateAlphaReward(_token, _account);
     pool.totalAlphaTokenReward = pool.totalAlphaTokenReward.sub(reward);
     sendAlphaReward(_account, reward);
-    setLatestReward(_token, _account);
+    setLatestAlphaMultiplier(_token, _account);
   }
 
+  /**
+   * @dev distribute Alpha tokens to the receiver
+   */
   function sendAlphaReward(address _recipient, uint256 _amount) internal {
     if (address(vestingAlpha) == address(0)) {
       distributor.alphaToken().transfer(_recipient, _amount);
@@ -1230,13 +1252,16 @@ contract LendingPool is Ownable, ILendingPool, IAlphaReceiver {
     }
   }
 
-  function setLatestReward(ERC20 _token, address _account) internal {
+  /**
+   * @dev set latest user's Alpha multiplier (borrower)
+   */
+  function setLatestAlphaMultiplier(ERC20 _token, address _account) internal {
     // No op if alpha distributor didn't be set in lending pool.
     if (address(distributor) == address(0)) {
       return;
     }
     Pool storage pool = pools[address(_token)];
     UserPoolData storage userData = userPoolData[_account][address(_token)];
-    userData.latestBorrowReward = userData.borrowShares.mul(pool.alphaRewardMultiplier).div(1e12);
+    userData.latestAlphaMultiplier = pool.alphaMultiplier;
   }
 }
